Restructure Brick.Events
This moves a lot of the logic out of Brick.Events, to make
it easier to isolate which module parts are shared, and which
parts are specific to individual event handlers.
